+++ adapters/Facebook/FacebookAdapter/OpenBidding/GADFBInterstitialRenderer.m
@@ -14,8 +14,8 @@
 
 #import "GADFBInterstitialRenderer.h"
 
-#import <FBAudienceNetwork/FBAudienceNetwork.h>
 #import <AdSupport/AdSupport.h>
+#import <FBAudienceNetwork/FBAudienceNetwork.h>
 
 #import "GADFBError.h"
 #import "GADMAdapterFacebookConstants.h"